Fix issue regarding to step does not receive environment variable
Environment variables exist but variable subtitution is not occuring due to
command is not being interpreted by a shell. (More detail at https://pkg.go.dev/os/exec)

So to archive the desired result in command, os.ExpandEnv is used on every
argument or flag value when command is being parsed.

In addition, simple jfrog plugin for deploying package to jfrog artifactory
also is added.

Signed-off-by: Loc Ngo <locngo@fortna.com>